fix: groupAlbum 에서 postmain 으로 파일 또는 폴더 이름 변경
