+++ src/pages/post/PostWrite.jsx
@@ -1,5 +1,9 @@
-import React from 'react';
+import React from 'react'
 
-export default function PostWrite() {
-  return <div>PostWrite</div>;
+function PostWrite() {
+  return (
+    <div>PostWrite</div>
+  )
 }
+
+export default PostWrite
